src/main/webapp/app/entities/my-cause/route/my-cause-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IMyCause, MyCause } from '../my-cause.model';
import { MyCauseService } from '../service/my-cause.service';

import { MyCauseRoutingResolveService } from './my-cause-routing-resolve.service';

describe('MyCause routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: MyCauseRoutingResolveService;
  let service: MyCauseService;
  let resultMyCause: IMyCause | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(MyCauseRoutingResolveService);
    service = TestBed.inject(MyCauseService);
    resultMyCause = undefined;
  });

  describe('resolve', () => {
    it('should return IMyCause returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultMyCause = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultMyCause).toEqual({ id: 123 });
    });

    it('should return new IMyCause if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultMyCause = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultMyCause).toEqual(new MyCause());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as MyCause })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultMyCause = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultMyCause).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
